add command to import partners type facilitator
