+++ lemarche/fixtures/tests/partners_import.csv
@@ -1,0 +1,4 @@
+FIRST_NAME,LAST_NAME,EMAIL,PHONE,POSITION
+Jean,Dupont,jean.dupont@facilitateur.fr,0123456789,Facilitateur clauses sociales
+Marie,Martin,marie.martin@accompagnement.fr,0987654321,Accompagnatrice insertion
+Pierre,Durand,pierre.durand@mediation.fr,0555666777,Médiateur social
