+++ lemarche/users/tests/test_commands.py
@@ -3,172 +3,18 @@ from io import StringIO
 from unittest.mock import patch
 
 from dateutil.relativedelta import relativedelta
-from django.contrib.messages import get_messages
 from django.core.management import call_command
 from django.core.validators import validate_email
 from django.db.models import F
 from django.test import TestCase, override_settings
-from django.urls import reverse
 
 from lemarche.companies.factories import CompanyFactory
 from lemarche.conversations.factories import TemplateTransactionalFactory
 from lemarche.conversations.models import TemplateTransactional, TemplateTransactionalSendLog
-from lemarche.favorites.factories import FavoriteListFactory
 from lemarche.siaes.factories import SiaeFactory
-from lemarche.tenders.factories import TenderFactory
-from lemarche.users import constants as user_constants
+from lemarche.users.constants import PARTNER_KIND_FACILITATOR
 from lemarche.users.factories import UserFactory
 from lemarche.users.models import User
-
-
-class UserModelTest(TestCase):
-    @classmethod
-    def setUpTestData(cls):
-        cls.user = UserFactory()
-
-    def test_str(self):
-        user = UserFactory(email="coucou@example.com")
-        self.assertEqual(str(user), "coucou@example.com")
-
-    def test_full_name(self):
-        user = UserFactory(first_name="Paul", last_name="Anploi")
-        self.assertEqual(user.full_name, "Paul Anploi")
-
-    def test_kind_detail_display(self):
-        user_siae = UserFactory(kind=user_constants.KIND_SIAE)
-        user_buyer = UserFactory(kind=user_constants.KIND_BUYER)
-        user_buyer_public = UserFactory(kind=user_constants.KIND_BUYER, buyer_kind=user_constants.BUYER_KIND_PUBLIC)
-        user_buyer_private_pme = UserFactory(
-            kind=user_constants.KIND_BUYER,
-            buyer_kind=user_constants.BUYER_KIND_PRIVATE,
-            buyer_kind_detail=user_constants.BUYER_KIND_DETAIL_PRIVATE_PME,
-        )
-        user_partner = UserFactory(kind=user_constants.KIND_PARTNER)
-        user_partner_facilitator = UserFactory(
-            kind=user_constants.KIND_PARTNER, partner_kind=user_constants.PARTNER_KIND_FACILITATOR
-        )
-        self.assertEqual(user_siae.kind_detail_display, "Structure")
-        self.assertEqual(user_buyer.kind_detail_display, "Acheteur")
-        self.assertEqual(user_partner.kind_detail_display, "Partenaire")
-        self.assertEqual(user_buyer_public.kind_detail_display, "Acheteur")
-        self.assertEqual(user_buyer_private_pme.kind_detail_display, "Acheteur : PME")
-        self.assertEqual(
-            user_partner_facilitator.kind_detail_display, "Partenaire : Facilitateur des clauses sociales"
-        )
-
-
-class UserModelQuerysetTest(TestCase):
-    @classmethod
-    def setUpTestData(cls):
-        cls.user = UserFactory()
-
-    def test_is_admin_bizdev(self):
-        UserFactory(kind=user_constants.KIND_ADMIN, is_staff=True)
-        UserFactory(kind=user_constants.KIND_ADMIN, is_staff=True, position="BizDev")
-        UserFactory(kind=user_constants.KIND_ADMIN, is_staff=True, position="Bizdev")
-        self.assertEqual(User.objects.count(), 1 + 3)
-        self.assertEqual(User.objects.is_admin_bizdev().count(), 2)
-
-    def test_has_company(self):
-        user_2 = UserFactory()
-        CompanyFactory(users=[user_2])
-        self.assertEqual(User.objects.count(), 1 + 1)
-        self.assertEqual(User.objects.has_company().count(), 1)
-
-    def test_has_siae(self):
-        user_2 = UserFactory()
-        SiaeFactory(users=[user_2])
-        self.assertEqual(User.objects.count(), 1 + 1)
-        self.assertEqual(User.objects.has_siae().count(), 1)
-
-    def test_has_tender(self):
-        user_2 = UserFactory()
-        TenderFactory(author=user_2)
-        self.assertEqual(User.objects.count(), 1 + 1)
-        self.assertEqual(User.objects.has_tender().count(), 1)
-
-    def test_has_favorite_list(self):
-        user_2 = UserFactory()
-        FavoriteListFactory(user=user_2)
-        self.assertEqual(User.objects.count(), 1 + 1)
-        self.assertEqual(User.objects.has_favorite_list().count(), 1)
-
-    def test_has_api_key(self):
-        UserFactory(api_key="coucou")
-        self.assertEqual(User.objects.count(), 1 + 1)
-        self.assertEqual(User.objects.has_api_key().count(), 1)
-
-    def test_has_email_domain(self):
-        UserFactory(email="test@ain.fr")
-        UserFactory(email="test@plateau-urbain.fr")
-        self.assertEqual(User.objects.count(), 1 + 2)
-        for EMAIL_DOMAIN in ["ain.fr", "@ain.fr"]:
-            with self.subTest(email_domain=EMAIL_DOMAIN):
-                self.assertEqual(User.objects.has_email_domain(email_domain=EMAIL_DOMAIN).count(), 1)
-
-    def test_with_siae_stats(self):
-        user_2 = UserFactory()
-        SiaeFactory(users=[user_2])
-        self.assertEqual(User.objects.count(), 1 + 1)
-        self.assertEqual(User.objects.with_siae_stats().filter(id=self.user.id).first().siae_count_annotated, 0)
-        self.assertEqual(User.objects.with_siae_stats().filter(id=user_2.id).first().siae_count_annotated, 1)
-
-    def test_with_tender_stats(self):
-        user_2 = UserFactory()
-        TenderFactory(author=user_2)
-        self.assertEqual(User.objects.count(), 1 + 1)
-        self.assertEqual(User.objects.with_tender_stats().filter(id=self.user.id).first().tender_count_annotated, 0)
-        self.assertEqual(User.objects.with_tender_stats().filter(id=user_2.id).first().tender_count_annotated, 1)
-
-    def test_chain_querysets(self):
-        user_2 = UserFactory(api_key="chain")
-        siae = SiaeFactory()
-        siae.users.add(user_2)
-        self.assertEqual(
-            User.objects.has_api_key().with_siae_stats().filter(id=user_2.id).first().siae_count_annotated, 1
-        )
-
-
-class UserModelSaveTest(TestCase):
-    def setUp(self):
-        pass
-
-    def test_update_last_updated_fields(self):
-        user = UserFactory()
-        self.assertIsNone(user.api_key)
-        self.assertIsNone(user.api_key_last_updated)
-        # new value: last_updated field will be set
-        user = User.objects.get(id=user.id)  # we need to fetch it again to pass through the __init__
-        user.api_key = "AZERTY"
-        user.save()
-        self.assertEqual(user.api_key, "AZERTY")
-        self.assertIsNotNone(user.api_key_last_updated)
-        api_key_last_updated = user.api_key_last_updated
-        # same value: last_updated field will not be updated
-        user = User.objects.get(id=user.id)
-        user.api_key = "AZERTY"
-        user.save()
-        self.assertEqual(user.api_key, "AZERTY")
-        self.assertEqual(user.api_key_last_updated, api_key_last_updated)
-        # updated value: last_updated field will be updated
-        user = User.objects.get(id=user.id)
-        user.api_key = "QWERTY"
-        user.save()
-        self.assertEqual(user.api_key, "QWERTY")
-        self.assertNotEqual(user.api_key_last_updated, api_key_last_updated)
-
-    def test_update_related_favorite_list_count_on_save(self):
-        user = UserFactory()
-        self.assertEqual(user.favorite_list_count, 0)
-        # create 2 lists
-        FavoriteListFactory(user=user)
-        FavoriteListFactory(user=user)
-        self.assertEqual(user.favorite_lists.count(), 2)
-        self.assertEqual(user.favorite_list_count, 2)
-        # delete 1 list
-        user.favorite_lists.first().delete()
-        self.assertEqual(user.favorite_lists.count(), 1)
-        self.assertEqual(user.favorite_list_count, 1)
 
 
 # To avoid different results when test will be run in the future, we patch
@@ -314,39 +160,6 @@ class UserAnonymizationTestCase(TestCase):
         self.assertFalse(TemplateTransactionalSendLog.objects.all())
 
 
-class UserAdminTestCase(TestCase):
-    def setUp(self):
-        UserFactory(is_staff=False, is_anonymized=False)
-        super_user = UserFactory(is_staff=True, is_superuser=True)
-        self.client.force_login(super_user)
-
-    def test_anonymize_action(self):
-        """Test the anonymize_users action from the admin"""
-
-        users_ids = User.objects.values_list("id", flat=True)
-        data = {
-            "action": "anonymize_users",
-            "_selected_action": users_ids,
-        }
-        # https://docs.djangoproject.com/en/5.1/ref/contrib/admin/#reversing-admin-urls
-        change_url = reverse("admin:users_user_changelist")
-        response = self.client.post(path=change_url, data=data)
-
-        self.assertEqual(response.status_code, 302)
-
-        data_confirm = {"user_id": users_ids}
-
-        # click on confirm after seeing the confirmation page
-        response_confirm = self.client.post(response.url, data=data_confirm)
-        self.assertEqual(response.status_code, 302)
-
-        self.assertTrue(User.objects.filter(is_staff=False).first().is_anonymized)
-        self.assertFalse(User.objects.filter(is_staff=True).first().is_anonymized)
-
-        messages_strings = [str(message) for message in get_messages(response_confirm.wsgi_request)]
-        self.assertIn("L'anonymisation s'est déroulée avec succès", messages_strings)
-
-
 class UserBuyerImportTestCase(TestCase):
     """Test the import_buyers management command"""
 
@@ -358,8 +171,8 @@ class UserBuyerImportTestCase(TestCase):
         call_command(
             "import_buyers",
             "lemarche/fixtures/tests/acheteurs_import.csv",
-            "grosse-banque",
             "USER_IMPORT_BUYERS_BANQUE",
+            "grosse-banque",
             5,
             stdout=StringIO(),  # avoid polluting the logs in test execution
         )
@@ -394,14 +207,14 @@ class UserBuyerImportTestCase(TestCase):
         with (
             patch("lemarche.users.management.commands.import_buyers.add_to_contact_list") as mock_add_to_contact_list,
             patch(
-                "lemarche.users.management.commands.import_buyers.send_new_user_password_reset_link"
+                "lemarche.users.management.commands.base_import_users.send_new_user_password_reset_link"
             ) as mock_send_new_user_password_reset_link,
         ):
             call_command(
                 "import_buyers",
                 "lemarche/fixtures/tests/acheteurs_import.csv",
-                "grosse-banque",
                 "USER_IMPORT_BUYERS_BANQUE",
+                "grosse-banque",
                 5,
                 stdout=out,
             )
@@ -418,8 +231,10 @@ class UserBuyerImportTestCase(TestCase):
         duplicated_user = User.objects.get(email="dupont.lajoie@camping.fr")
         self.assertEqual(duplicated_user.company, self.company)
         self.assertEqual(duplicated_user.company_name, self.company.name)
-        self.assertIn("L'acheteur dupont.lajoie@camping.fr est déjà inscrit, entreprise mise à jour.", out.getvalue())
-        self.assertIn("L'acheteur françois.perrin@celc.test.fr a été inscrit avec succès.", out.getvalue())
+        self.assertIn(
+            "L'utilisateur dupont.lajoie@camping.fr est déjà inscrit, informations mises à jour.", out.getvalue()
+        )
+        self.assertIn("L'utilisateur françois.perrin@celc.test.fr a été inscrit avec succès.", out.getvalue())
 
     def test_email_dns_added(self):
         """Ensure that email domains are added to the company's email_domain_list, without duplicates."""
@@ -427,8 +242,8 @@ class UserBuyerImportTestCase(TestCase):
         call_command(
             "import_buyers",
             "lemarche/fixtures/tests/acheteurs_import.csv",
-            "grosse-banque",
             "USER_IMPORT_BUYERS_BANQUE",
+            "grosse-banque",
             5,
             stdout=StringIO(),
         )
@@ -439,8 +254,8 @@ class UserBuyerImportTestCase(TestCase):
         call_command(
             "import_buyers",
             "lemarche/fixtures/tests/acheteurs_import.csv",
-            "grosse-banque",
             "USER_IMPORT_BUYERS_BANQUE",
+            "grosse-banque",
             5,
             stdout=StringIO(),
         )
@@ -463,8 +278,8 @@ class UserBuyerImportTestCase(TestCase):
             call_command(
                 "import_buyers",
                 "lemarche/fixtures/tests/acheteurs_import.csv",
-                "grosse-banque",
                 "USER_IMPORT_BUYERS_BANQUE",
+                "grosse-banque",
                 5,
                 stdout=StringIO(),
             )
@@ -479,3 +294,79 @@ class UserBuyerImportTestCase(TestCase):
 
         self.company.refresh_from_db()
         self.assertEqual(self.company.email_domain_list, ["celc.test.fr"])
+
+
+class UserPartnerImportTestCase(TestCase):
+    """Test the import_partners management command"""
+
+    def setUp(self):
+        TemplateTransactionalFactory(code="USER_IMPORT_PARTNERS_FACILITATOR")
+
+    def test_import_partner(self):
+        with patch(
+            "lemarche.users.management.commands.base_import_users.send_new_user_password_reset_link"
+        ) as mock_send_new_user_password_reset_link:
+
+            call_command(
+                "import_partners",
+                "lemarche/fixtures/tests/partners_import.csv",
+                "USER_IMPORT_PARTNERS_FACILITATOR",
+                stdout=StringIO(),
+            )
+
+        self.assertEqual(mock_send_new_user_password_reset_link.call_count, 3)
+
+        self.assertQuerySetEqual(
+            User.objects.all(),
+            [
+                "<User: jean.dupont@facilitateur.fr>",
+                "<User: marie.martin@accompagnement.fr>",
+                "<User: pierre.durand@mediation.fr>",
+            ],
+            ordered=False,
+            transform=lambda x: repr(x),
+        )
+
+        self.assertQuerySetEqual(
+            User.objects.filter(partner_kind=PARTNER_KIND_FACILITATOR),
+            [
+                "<User: jean.dupont@facilitateur.fr>",
+                "<User: marie.martin@accompagnement.fr>",
+                "<User: pierre.durand@mediation.fr>",
+            ],
+            ordered=False,
+            transform=lambda x: repr(x),
+        )
+
+    def test_already_registered_partner(self):
+        """
+        Existing partner should not have its information updated and an error should be logged
+        """
+        UserFactory(email="jean.dupont@facilitateur.fr", kind=User.KIND_BUYER)
+        out = StringIO()
+        with patch(
+            "lemarche.users.management.commands.base_import_users.send_new_user_password_reset_link"
+        ) as mock_send_new_user_password_reset_link:
+            call_command(
+                "import_partners",
+                "lemarche/fixtures/tests/partners_import.csv",
+                "USER_IMPORT_PARTNERS_FACILITATOR",
+                stdout=out,
+            )
+            self.assertEqual(mock_send_new_user_password_reset_link.call_count, 2)
+
+        self.assertQuerySetEqual(
+            User.objects.all(),
+            [
+                "<User: jean.dupont@facilitateur.fr>",
+                "<User: marie.martin@accompagnement.fr>",
+                "<User: pierre.durand@mediation.fr>",
+            ],
+            ordered=False,
+            transform=lambda x: repr(x),
+        )
+        self.assertIn(
+            "L'utilisateur jean.dupont@facilitateur.fr est déjà inscrit, aucune mise à jour.", out.getvalue()
+        )
+        self.assertIn("L'utilisateur marie.martin@accompagnement.fr a été inscrit avec succès.", out.getvalue())
+        self.assertIn("L'utilisateur pierre.durand@mediation.fr a été inscrit avec succès.", out.getvalue())
